Created Spider class
