+++ Houses.csproj
@@ -63,6 +63,7 @@
     <Compile Include="Home.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Spider.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
